Refactor lifecycle waitUntil tests into one file
