+++ tests/static/prolonged_network_idle.html
@@ -3,23 +3,23 @@
 <head></head>
 
 <body>
-    <div id="prolongNetworkIdleLoad">Waiting...</div>
-    <div id="serverMsg">Waiting...</div>
+    <div id="pingRequestText">Waiting...</div>
+    <div id="pingJSText">Waiting...</div>
 
     <script>
-        var prolongNetworkIdleLoadOutput = document.getElementById("prolongNetworkIdleLoad");
+        var pingRequestTextOutput = document.getElementById("pingRequestText");
 
-        var p = prolongNetworkIdleLoad();
+        var p = pingRequestText();
         p.then(() => {
-            prolongNetworkIdleLoadOutput.innerText += ' - for loop complete';
+            pingRequestTextOutput.innerText += ' - for loop complete';
         })
 
-        async function prolongNetworkIdleLoad() {
+        async function pingRequestText() {
             for (var i = 0; i < 4; i++) {
                 await fetch('/ping')
                     .then(response => response.text())
                     .then((data) => {
-                        prolongNetworkIdleLoadOutput.innerText = 'Waiting... ' + data;
+                        pingRequestTextOutput.innerText = 'Waiting... ' + data;
                     });
             }
         }
